requirements.txt
streamlit
pandas
python-dateutil
psycopg2-binary
python-dotenv
